Implement team byes in draws
This commit implements teams being marked as being given a bye for a
round. To support this, placeholder debates and ballots are created for
them, which bypass results. Rather, they are treated as not a debate
when getting their ballot, and links to them are hidden. In draw tables,
their situation is highlighted, and by default moved to the bottom of
the table.

The creation of bye teams and debates is handled by the draw manager,
which is now given a couple more options from the (draw rules)
preferences for how to select byes and how to treat them:
- Bye selection: How teams are chosen to get the bye
  - Random: Any team could get the bye,
  - Lowest: The lowest-ranking team (if supported by the draw) is
            excluded.
- Bye results: What type of results would we give bye teams
  - None: Skip creating ballots for them, equivalent to a loss
  - Points: Give them a winning TeamScore without speaks

As to not tank bye teams within their bracket, average speaks should be
used rather than sum of speaks as a tiebreaker.